update aiflow download file drive

+++ dags/misa_nhom_sp.py
@@ -16,6 +16,7 @@ TEMP_PATH = Variable.get("temp_path")
 ## Conection
 HOOK_MSSQL = Variable.get("mssql_connection")
 
+FOLDER_NAME = 'nhomsanpham'
 
 default_args = {
     "owner": "hieulc",
@@ -35,7 +36,7 @@ default_args = {
 def Misa_group_products():
     @task
     def remove_files():
-        folder = os.path.join(TEMP_PATH , 'nhomsanpham')
+        folder = os.path.join(TEMP_PATH , FOLDER_NAME)
         files = glob.glob(os.path.join(folder, '*'))
         for i in files:
             try:
@@ -46,51 +47,52 @@ def Misa_group_products():
     ######################################### API ################################################
     @task
     def download_latest_file() -> str:
-        folder_name = 'nhomsanpham'
         folder_id = Variable.get("folder_id_sanpham_nhom")
-        return download_file_drive(folder_name=folder_name, folder_id=folder_id)
+        return download_file_drive(folder_name=FOLDER_NAME, folder_id=folder_id)
 
     ######################################### INSERT DATA ################################################
     @task
-    def insert_group_products(local_file: str) -> None:
+    def insert_group_products(list_local_file: list) -> None:
         hook = mssql.MsSqlHook(HOOK_MSSQL)
         sql_conn = hook.get_conn()
         cursor = sql_conn.cursor()
-        df = pd.read_excel(local_file, index_col=None, engine='openpyxl')
-        print(df.columns)
-        sql_del = f"delete from [dbo].[3rd_misa_group_products] where [ma_hang] in {tuple([str(i) for i in df['MÃ HÀNG HOÁ'].tolist()])};"
-        print(sql_del)
-        cursor.execute(sql_del)
-        values = []
-        if len(df) > 0:
-            sql = """
-                    INSERT INTO [dbo].[3rd_misa_group_products](
-                        [ma_hang]
-                        ,[ten_hang]
-                        ,[nhan_hang]
-                        ,[nhom_nhan_hang]
-                        ,[dtm_creation_date])
-                    VALUES(%s, %s, %s, %s, getdate())
-                """
-            for _index, row in df.iterrows():
-                value = (
-                            str(row[0]),
-                            str(row[1]),
-                            str(row[2]),
-                            str(row[3]),
-                )
-                values.append(value)
-            cursor.executemany(sql, values)
+        if len(list_local_file) > 0:
+            for local_file in list_local_file:
+                df = pd.read_excel(local_file, index_col=None, engine='openpyxl')
+                print(df.columns)
+                sql_del = f"delete from [dbo].[3rd_misa_group_products] where [ma_hang] in {tuple([str(i) for i in df['MÃ HÀNG HOÁ'].tolist()])};"
+                print(sql_del)
+                cursor.execute(sql_del)
+                values = []
+                if len(df) > 0:
+                    sql = """
+                            INSERT INTO [dbo].[3rd_misa_group_products](
+                                [ma_hang]
+                                ,[ten_hang]
+                                ,[nhan_hang]
+                                ,[nhom_nhan_hang]
+                                ,[dtm_creation_date])
+                            VALUES(%s, %s, %s, %s, getdate())
+                        """
+                    for _index, row in df.iterrows():
+                        value = (
+                                    str(row[0]),
+                                    str(row[1]),
+                                    str(row[2]),
+                                    str(row[3]),
+                        )
+                        values.append(value)
+                    cursor.executemany(sql, values)
 
-        print(f"Inserted {len(values)} rows in database with {df.shape[0]} rows")
-        sql_conn.commit()
+                print(f"Inserted {len(values)} rows in database with {df.shape[0]} rows")
+                sql_conn.commit()
         sql_conn.close()
 
     ############ DAG FLOW ############
 
-    local_file = download_latest_file()
-    insert_task = insert_group_products(local_file)
-    remove_files() >> local_file >> insert_task
+    list_local_file = download_latest_file()
+    insert_task = insert_group_products(list_local_file)
+    remove_files() >> list_local_file >> insert_task
     # remove_files() >> insert_task
